pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "videoqa-gate"
version = "0.1.0"
description = "Three-layer quality gate + evaluation-correction loop for AI-generated short videos"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["video", "quality-assessment", "ai-video", "evaluation", "self-correction"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "ffmpeg-python>=0.2.0",
    "imagehash>=4.3",
    "pysrt>=1.1",
    "pyyaml>=6.0",
    "numpy>=1.24",
    "Pillow>=10.0",
]

[project.optional-dependencies]
ml = [
    "open-clip-torch>=2.24",
    "torch>=2.1",
    "opencv-python>=4.8",
    "scenedetect[opencv]>=0.6",
]
vlm = [
    "google-generativeai>=0.8",
]
pipeline = [
    "langgraph>=0.2",
    "moviepy>=2.0",
]
all = [
    "videoqa-gate[ml,vlm,pipeline]",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
videoqa = "src.quality_gate:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py310"
line-length = 100
